+++ frontend/guest-app/src/components/Question/QuestionCard/QuestionCardHeader.js
@@ -1,10 +1,11 @@
 import React from "react";
 import styled from "styled-components";
 import Grid from "@material-ui/core/Grid";
-import {Typography} from "@material-ui/core";
 import UserAvatar from "../../UserAvatar/UserAvatar.js";
 import LikeButton from "../../LikeButton/LikeButton.js";
 import useLikeButton from "../../LikeButton/useLikeButton.js";
+import QuestionCardDate from "./QuestionsCardDate.js";
+import QuestionUserName from "./QuestionCardUserName.js";
 
 const QuestionInfo = styled.div`
 	display: flex;
@@ -12,26 +13,6 @@ const QuestionInfo = styled.div`
 	align-content: left;
 	margin-left: 1rem;
 `;
-
-function QuestionUserName({userName}) {
-	return (
-		<Typography
-			color={"textPrimary"}
-			variant={"subtitle1"}
-			style={{fontWeight: "bold"}}
-		>
-			{userName}
-		</Typography>
-	);
-}
-
-function QuestionDate({date}) {
-	return (
-		<Typography color={"textSecondary"} variant={"body1"}>
-			{new Date(parseInt(date, 10)).toLocaleString()}
-		</Typography>
-	);
-}
 
 const QuestionHeaderStyle = styled.div`
 	display: flex;
@@ -51,7 +32,7 @@ function QuestionHeader(props) {
 				<UserAvatar userName={guestName} />
 				<QuestionInfo>
 					<QuestionUserName userName={guestName} />
-					<QuestionDate date={createdAt} />
+					<QuestionCardDate date={createdAt} />
 				</QuestionInfo>
 			</Grid>
 			<LikeButton {...likeButton} />
