+++ frontend/guest-app/src/components/Question/QuestionCard/QuestionCardUserName.js
@@ -1,0 +1,14 @@
+import {Typography} from "@material-ui/core";
+import React from "react";
+
+export default function QuestionUserName({userName}) {
+	return (
+		<Typography
+			color={"textPrimary"}
+			variant={"subtitle1"}
+			style={{fontWeight: "bold"}}
+		>
+			{userName}
+		</Typography>
+	);
+}
